+++ src/main/java/com/sicca/repository/ParametroCultivoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ParametroCultivoRepository extends JpaRepository<ParametroCultivoEntity, Long> {
+public interface ParametroCultivoRepository extends JpaRepository<ParametroCultivoEntity, Integer> {
 }
