+++ src/main/java/com/sicca/repository/TipoSensorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoSensorRepository extends JpaRepository<TipoSensorEntity, Long> {
+public interface TipoSensorRepository extends JpaRepository<TipoSensorEntity, Integer> {
 }
